requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10

# Scheduling
schedule>=1.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# HTTP requests
requests>=2.28.0
httpx>=0.24.0

# Data validation
pydantic>=2.0.0

# Logging
structlog>=23.0.0

# Environment variables
python-dotenv>=1.0.0

# Bluesky AT Protocol
atproto>=0.0.39

# Docker integration testing
docker>=6.0.0

# Management API
flask>=2.3.0